Add dynamic inventory script for ganeti clusters
This will query the cluster and return the nodes and instances (aka
guests) in the cluster.  This will create two groups:  nodes and
instances.  You can use the configuration file (ganeti.ini in the same
directory as the script) to set the RAPI host to query or you can use
the environment variable GNT_RAPI_HOST and GNT_RAPI_PORT to determine
which ganeti cluster to configure.